% Generated by roxygen2 (4.0.2): do not edit by hand
\name{betaCor}
\alias{betaCor}
\title{Compute within-sample correlations of pairs of beta-values.}
\usage{
betaCor(methpat, pair_type = c("neighbours", "all"), mtuples,
  max_ipd = 2000, method = c("pearson", "spearman"), feature,
  same_feature = FALSE, feature_name)
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object containing 1-tuples.}

\item{pair_type}{A character string giving the type of pairs to be
constructed when computing correlations. One of "\code{neighbours}" or "\code{all}", can be abbreviated. Please see the below section, "Construction
of pairs of beta-values", for details.}

\item{mtuples}{An \code{\link{MTuples}} object with the locations of all
methylation loci 1-tuples in the sample/reference genome. Only required if
\code{pair_type = "neighbours"} and otherwise ignored. Please see the below
section, "Construction of pairs of beta-values", for details.}

\item{max_ipd}{An \code{integer} specifying the maximal IPD of pairs
(default: 2000). Only required if \code{pair_type = "all"} and otherwise
ignored. Please see the below section, "Construction of pairs of
beta-values", for details.}

\item{method}{A character string indicating which correlation coefficient is
to be computed. One of "\code{pearson}" (default) or "\code{spearman}", can
be abbreviated.}

\item{feature}{An optional \code{\link[GenomicRanges]{GRanges}} object with
the locations of a "genomic feature". This
\code{\link[GenomicRanges]{GRanges}} object must be disjoint (see
\code{\link[GenomicRanges]{isDisjoint}}). The \code{feature}Please see the
below section, "Stratifying pairs by a genomic feature", for details.}

\item{same_feature}{A \code{logical(1)}. When stratifying by \code{feature},
should a pair be required to be in the "same element" of the feature. Please
see the below section, "Stratifying pairs by a genomic feature", for details.}

\item{feature_name}{A character string with the name of the feature, which
is used in the output, e.g., "CpG island".}
}
\description{
Given a \code{\link{MethPat}} object containing 1-tuples, \code{betaCor}
computes the within-sample correlations of pairs of beta-values. The manner
in which the pairs are constructed is determined by additional arguments -
see the section, "Construction of pairs of beta-values". Correlations are
stratified by the \code{strand}, the intra-pair distance and the
\code{feature} (if supplied) of the pairs.
}
\section{Constructing pairs of beta-values}{

There are two algorithms for constructing pairs of beta-values:
\code{neighbours}, which requires the specification of \code{mtuples}, and
\code{all}, which requires the specification of \code{max_ipd}.

\itemize{
 \item{\code{pair_type = "neighbours"}:}{ Creates pairs of beta-values
 from neighbouring methylation loci. It checks that the constructed pairs
 are neighbours by comparing these to the set of all known methylation loci
 in the sample, hence this must be given by \code{mtuples}.}

 \item{\code{pair_type = "all"}:}{ Creates pairs of beta-values using all
 methylation loci in the sample such that each pair has an intra-pair
 distance less than or equal to \code{max_ipd}.}

}
}

\section{Stratifying pairs by a genomic feature}{

Pairs of beta-values may be stratified by whether they are inside
or outside of a "genomic feature". For our example, we will use CpG islands
as our feature. In this case \code{feature} should be a
\code{\link[GenomicRanges]{GRanges}} object containing all CpG islands in
the genome.

The assignment of pairs as "inside" or "outside" is suprisingly nuanced. The
\code{same_feature} argument controls how a pair is assigned
as being "inside" or "outside" the feature.

Please think carefully about what defines a pair as being "inside" and
"outside" a feature for your particular needs. While I have attempted to
make \code{betaCor} fairly general, if you have more complex
requirements then you may find the available options to be insufficient.

\itemize{
 \item{\code{same_feature = FALSE}: }{A pair is considered to be "inside"
 the feature if and only if both loci that make up in the pair overlap an
 element of \code{feature}. However, the two loci in each pair may overlap
 different elements of \code{feature}, e.g., different CpG islands.
 Similarly, a pair is considered to be "outside" the feature if and only if
 both loci that make up the pair overlap a "gap" between elements of
 \code{feature}. However, the two loci in each pair may overlap different
 "gaps", non-CpG island regions of the genome. This is the default.
 }
 \item{\code{same_feature = TRUE}: A pair is considered to be "inside" the
 feature only if and only if both loci that make up the pair are in the same
 feature, e.g., the same CpG island. Similarly, a pair is considered
 "outside" the feature only if and only if both loci that make up the pair
 are in the same "gap" between elements of \code{feature}. All other pairs
  are discarded, e.g., those where both loci that make up the pair are in a
  CpG island but lie in distinct CpG islands.}
}
}

