% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{betaVal,MethPat-method}
\alias{betaVal,MethPat-method}
\title{Compute beta-values.}
\usage{
\S4method{betaVal}{MethPat}(x, min_cov = 0L)
}
\arguments{
\item{x}{A \code{\link{MethPat}} object containing 1-tuples.}

\item{min_cov}{An \code{integer} specifying the minimum coverage required
in order order to compute a beta-value. Samples/sites with coverage less
than \code{min_cov} will have the corresponding beta-value set to \code{NA}.}
}
\value{
A \code{\link[base]{matrix}}, with the same dimensions and dimension
names as \code{x}, of beta-values at each methylation loci in each sample.
}
\description{
Compute beta-values.
}

