% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collapseStrand.R
\name{collapseStrand}
\alias{collapseStrand}
\title{Collapse a \code{\link{MethPat}} object by strand.}
\usage{
collapseStrand(methpat)
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object.}
}
\value{
An updated version of the \code{\link{MethPat}} object, where the
loci and their associated counts have been combined across strands.
}
\description{
A \code{\link{MethPat}} object may contain separate counts for the positive
(\code{+}) and negative (\code{-}) strands. In some instances it may be
desirable to collapse these observations across strands. This function will
combine the loci and sum their associated counts. The co-ordinates of the
collapsed loci will have the \code{\link[GenomicTuples]{strand}} set to
\code{*}, although positions will be with respect to the forward strand. For
example, the CpG with co-ordinates \code{chr1:10} on the \code{+} strand and
\code{chr1:11} on the \code{-} strand will be collapsed to \code{chr1:10}
with the strand set to \code{*}.
}
\details{
\strong{NOTE}: Collapsing by strand is only possible if the
\code{\link{methtype}} of the object is \code{CG}, since neither \code{CHG}
nor \code{CHH} is a strand-symmetric motif.

\strong{NOTE}: Any \code{mcols} of the \code{methpat} object will be dropped.
}

