% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cometh}
\alias{cometh}
\title{Estimate within-sample, within-fragment co-methylation using 2-tuples.}
\usage{
cometh(methpat, pair_type = c("all", "strict_adjacent"), ref_loci,
  method = c("lor", "pearson"), min_cov = 5L, alternative = c("two.sided",
  "less", "greater"), conf.level = 0.95, feature, offset = 1)
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object containing 2-tuples.}

\item{pair_type}{A character string giving the type of pairs to be
constructed when computing correlations. One of "\code{neighbours}" or
"\code{all}", can be abbreviated. Please see the below section, "Choice of
'pair_type'", for details.}

\item{ref_loci}{An \code{\link{MTuples}} object with the locations of all
methylation loci 1-tuples in the sample/reference genome. Only required if
\code{pair_type = "strict_adjacent"} and otherwise ignored. Please see the below
section, "Choice of 'pair_type'", for details.}

\item{method}{A character string giving the method used to estimate
co-methylation. One of "\code{lor}" or "\code{pearson}", can be abbreviated.
Please see the below section, "Choice of statistic", for details.}

\item{min_cov}{}
}
\description{
Estimate within-sample, within-fragment co-methylation using 2-tuples.
}
\section{Choice of 'pair_type'}{

}

\section{Choice of statistic}{

}

