% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MBias.R
\docType{methods}
\name{filter,MBias-method}
\alias{filter,MBias-method}
\title{Filter MBias}
\usage{
\S4method{filter}{MBias}(object, threshold = 3, methylationType = c("CG",
  "CHG", "CHG"))
}
\arguments{
\item{object}{An \code{\link{MBiase}} object.}

\item{threshold}{The threshold at which to filter.}

\item{methylationType}{The type of methylation to filter.}
}
\value{
A \code{data.table} with the sample name (\code{sampleName}), read
(\code{R1} for read-1 and \code{R2} for read-2), methylation type
(\code{methylationType}) and positions to filter out (\code{irp} = ignore
read positions). The value in the \code{irp} column is designed so that it
can be copy-pasted as an argument to \code{methtuple}.
}
\description{
For each sample, this identifies all read-positions with a normalise M-bias
greater than the threshold.
}

