% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_out_variants}
\alias{filter_out_variants}
\title{Filter out variants from \code{MethPat} object.}
\usage{
filter_out_variants(methpat, variant_files, remove = FALSE,
  verbose = getOption("verbose"), bpparam = bpparam())
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object.}

\item{variant_files}{A named \code{character} vector with the paths of the
\code{VCF}s containing the variant calls. Currently only VCFs created by
\code{Bis-SNP} are supported and it is required that there is one \code{VCF}
per sample and that the order of \code{variant_files} corresponds to the
samples in \code{methpat} (i.e.,
\code{identical(names(variant_files), colnames(methpat)} is \code{TRUE}).}

\item{remove}{A \code{logical(0)}. If \code{TRUE} rows of the \code{MethPat} object
where all samples have a variant are removed from the object entirely,
otherwise these are retained, albeit with the assay counts set to \code{NA}.}

\item{verbose}{A \code{\link{logical(1)}} indicating whether progress
should be reported via calls to \code{\link[base]{message}}.}

\item{bpparam}{A \code{\link[BiocParallel]{bpparam}} object specifying the
parallelisation strategy, if any. See below for a discussion of
parallelisation options available with \code{read.methtuple}.}
}
\value{
An updated version of the \code{\link{MethPat}} object, where
variants have had their corresponding assays counts set to \code{NA}. Tuples
where all samples are \code{NA} are retained (unless \code{remove} is
\code{TRUE}).
}
\description{
Filter out variants from \code{MethPat} object.
}
\section{Parallelisation}{

Parallelisation of \code{filter_out_variants} is partially supported.
\code{VCF}s are read-in and processed in parallel, where appropriate.
Parallelisation uses the \pkg{BiocParallel} package. By default this uses a
\code{\link[BiocParallel]{MulticoreParam()}} instance or the user’s
preferred back-end if they have used \code{\link[BiocParallel]{register}}.
Please consult the \pkg{BiocParallel} documentation for details on
registering a parallel backend and parallelisation support available on
different operating systems.
}
\examples{
\dontrun{
## TODO
}
}

