% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_variants}
\alias{filter_variants}
\title{Filter out variants from MethPat object.}
\usage{
filter_variants(methpat, variant_files, remove = FALSE,
  param = ScanVcfParam(fixed = c("FILTER", ), info = c("CS")))
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object.}

\item{variant_files}{A \code{character} vector with the names of the VCFs
containing the variant calls. Currently only VCFs created by \code{Bis-SNP}
are supported and it is required that there is one VCF per sample and that
the order of \code{variant_files} corresponds to the samples in
\code{methpat}.}

\item{A}{\code{logical(0)}. If \code{TRUE} rows of the \code{MethPat} object
where all samples have a variant are removed from the object entirely,
otherwise these are retained, albeit with the assay counts set to \code{NA}.}
}
\value{
An updated version of the \code{\link{MethPat}} object, where
variants have had their corresponding assays counts set to \code{NA}. Tuples
where all samples are \code{NA} are retained (unless \code{remove} is
\code{TRUE}).
}
\description{
Filter out variants from MethPat object.
}

