% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MethInfo-class}
\alias{MethInfo-class}
\title{MethInfo objects}
\description{
A \code{MethInfo} object is an object that contains basic
information about a set of methylation loci. Currently the only attributes
are the type or context of the methylation loci, but more attributes might
be added in the future as the need arises.
}
\details{
Typically \code{MethInfo} objects are not used directly but are part of
higher level objects. Those higher level objects will generally provide a
\code{methinfo} accessor for getting/setting their \code{MethInfo} component.
}
\section{Constructor}{

\describe{
\item{}{
 \code{MethInfo(methtype)}: Creates a \code{MethInfo} object.
 }
}
}

\section{Accessor methods}{

In the code snippets below, \code{x} is a \code{MethInfo} object.
 \describe{
   \item{}{
     \code{methtype(x)}, \code{methtype(x) <- value}:
     Get/set the methylation type of \code{x}. \code{value} must be a
     character vector: \code{"CG"} (\emph{i.e.}, CpG), \code{"CHG"},
     \code{"CHH"}, \code{"CNN"} or some combination of these, e.g.,
     \code{c("CG", "CHG")} (\code{NA_character_} is also allowed,
     but not recommended).
   }
 }
}

\section{Combining MethInfo objects}{

There is no \code{c} method for \code{MethInfo} objects. Rather, a
\code{merge} method is provided.

In the code snippet below, \code{x} and \code{y} are \code{MethInfo} objects.
 \describe{
   \item{}{
     \code{merge(x, y)}:
     Merge \code{x} and \code{y} into a single \code{MethInfo} object where
     the methylation type is the union of \code{methtype(x)} and
     \code{methtype(y)}. If the \code{methtype} of any object is missing
     (\code{NA_character_}) then the merged \code{methtype} is also missing
     (\code{NA_character}).
   }
 }
}
\examples{
x <- MethInfo("CG")
 y <- MethInfo(c("CHG", "CG"))
 y # NB: Pretty-prints methylation type as "CG/CHG"
 methtype(y) # NB: Returns the methylation type as a character vector
 merge(x, y)
}
\author{
Peter Hickey
}

