% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{methinfo}
\alias{methinfo}
\alias{methinfo<-}
\alias{methtype}
\alias{methtype<-}
\title{Accessing/modifying methylation information}
\usage{
methinfo(x)

methinfo(x) <- value

methtype(x)

methtype(x) <- value
}
\arguments{
\item{x}{The object from/on which to get/set the methylation information.}

\item{value}{A character vector of methylation types: \code{"CG"} 
(\emph{i.e.}, CpG), \code{"CHG"}, \code{"CHH"}, \code{"CN"} or some 
combination of these, e.g., \code{c("CG", "CHG")} (\code{NA_character_} is 
also allowed, but not recommended).}
}
\description{
A set of generic functions for getting/setting/modifying the methylation 
information stored in an object.
}
\details{
The \linkS4class{MethInfo} class plays a central role for the functions 
described in this man page. For classes that implement it, \code{methinfo} 
should return a \linkS4class{MethInfo} object. Examples of containers that 
have a \code{methinfo} gettter and setter: \linkS4class{MTuples} and 
\linkS4class{MTuplesList} classes.
}

