% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethPat-class.R
\docType{methods}
\name{methLevel,MethPat-method}
\alias{methLevel}
\alias{methLevel,MethPat-method}
\title{Compute methylation levels.}
\usage{
\S4method{methLevel}{MethPat}(object, statistic = c("beta-values",
  "M-values"), min_cov = 1L, offset = 1L)
}
\arguments{
\item{statistic}{A \code{character} string indicating which methylation
level statistic is to be computed. One of "\code{beta-values}" or
"\code{M-values}" (see below).}

\item{min_cov}{An \code{integer} specifying the minimum coverage required
in order to compute a beta-value. Samples/sites with coverage less
than \code{min_cov} will have the corresponding methylation level set to
\code{NA}.}

\item{offset}{A \code{numeric} vector with length 1 used when computing
M-values (default: 1).}

\item{x}{A \code{\link{MethPat}} object containing 1-tuples.}
}
\value{
A \code{\link[base]{matrix}}, with the same dimensions and dimension
names as \code{x}, of methylation levels at each methylation loci in each
sample.
}
\description{
Compute methylation levels.
}
\details{
TODO: Define beta-values and M-values. Note any differences with how others
define beta-values or M-values, e.g., minfi.
}

