% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationTuples.R
\docType{package}
\name{MethylationTuples-package}
\alias{MethylationTuples-package}
\title{Tools for managing, analysing, and  visualising methylation patterns at 
genomic tuples
\pkg{MethylationTuples} provides tools for managing, analysing, and 
visualising methylation patterns at genomic tuples. These include analyses 
of co-methylation and epipolymorphism.}
\description{
Please refer to the vignettes to see how to use the \pkg{MethylationTuples}
package.
}

