% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTuplesList-class.R
\docType{class}
\name{MTuplesList-class}
\alias{MTuplesList}
\alias{MTuplesList-class}
\alias{methinfo,MTuplesList-method}
\alias{methinfo<-,MTuplesList,MethInfo-method}
\alias{methtype,MTuplesList-method}
\alias{methtype<-,MTuplesList,character-method}
\alias{relistToClass,MTuples-method}
\alias{show,MTuplesList-method}
\title{MTuplesList objects}
\usage{
MTuplesList(...)

\S4method{methinfo}{MTuplesList}(x)

\S4method{methtype}{MTuplesList}(x)

\S4method{methinfo}{MTuplesList,MethInfo}(x) <- value

\S4method{methtype}{MTuplesList,character}(x) <- value

\S4method{relistToClass}{MTuples}(x)

\S4method{show}{MTuplesList}(object)
}
\arguments{
\item{x}{The object from/on which to get/set the methylation information.}

\item{value}{A character vector of methylation types: \code{"CG"} 
(\emph{i.e.}, CpG), \code{"CHG"}, \code{"CHH"}, \code{"CN"} or some 
combination of these, e.g., \code{c("CG", "CHG")} (\code{NA_character_} is 
also allowed, but not recommended).}

\item{object}{An MTuplesList object.}

\item{...}{\code{\link{MTuples}} objects. All must contain the same 
\code{size} tuples.}
}
\description{
The \code{MTuplesList} class is a container for storing a collection of 
\code{\link{MTuples}} objects. The \code{MTuplesList} class is almost 
identical to the \code{\link[GenomicTuples]{GTuplesList}} on which it is 
based.
}
\details{
Please see \linkS4class{GTuplesList} for a description of available 
methods. The only additional methods are \code{methinfo} and 
\code{methtype}, which are identical to their \linkS4class{MTuples}
counterparts.
}
\examples{
## TODO

}
\seealso{
\linkS4class{MTuples}, \linkS4class{GTuplesList}
}

