% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/patternFreqs.R
\name{patternFreqs}
\alias{patternFreqs}
\title{A naive estimate of the frequency of methylation patterns.}
\usage{
patternFreqs(methpat, min_cov = 5L, ...)
}
\description{
A naive estimate of the frequency of methylation patterns.
}

