% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MBias.R
\name{readMBias}
\alias{readMBias}
\title{Read Bismark's M-bias file.}
\usage{
readMBias(file, sampleName)
}
\arguments{
\item{file}{A Bismark M-bias file (with \code{.M-bias.txt} extension).}

\item{sampleName}{The name of the sample. Will guess based on \code{file}
if argument is missing.}
}
\value{
A \code{\link{MBias}} object.
}
\description{
Read Bismark's M-bias file.
}

