% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SparseMethPat-class.R
\docType{class}
\name{SparseMethPat-class}
\alias{IPD,SparseMethPat-method}
\alias{SparseMethPat}
\alias{SparseMethPat-class}
\alias{methinfo,SparseMethPat-method}
\alias{methinfo<-,SparseMethPat,ANY-method}
\alias{methtype,SparseMethPat-method}
\alias{methtype<-,SparseMethPat,ANY-method}
\alias{rowTuples,SparseMethPat-method}
\alias{rowTuples<-,SparseMethPat-method}
\alias{size,SparseMethPat-method}
\alias{tuples,SparseMethPat-method}
\alias{tuples<-,SparseMethPat-method}
\title{SparseMethPat objects}
\usage{
SparseMethPat(sparseAssays = sparseAssays(), rowTuples = MTuples(),
  colData = DataFrame(), metadata = list())

\S4method{rowTuples}{SparseMethPat}(x, ...)

\S4method{methinfo}{SparseMethPat}(x)

\S4method{methtype}{SparseMethPat}(x)

\S4method{rowTuples}{SparseMethPat}(x, ...) <- value

\S4method{methinfo}{SparseMethPat,ANY}(x) <- value

\S4method{methtype}{SparseMethPat,ANY}(x) <- value

\S4method{size}{SparseMethPat}(x)

\S4method{tuples}{SparseMethPat}(x)

\S4method{tuples}{SparseMethPat}(x) <- value

\S4method{IPD}{SparseMethPat}(x)
}
\description{
SparseMethPat objects
}

