% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{strandCollapse}
\alias{strandCollapse}
\title{Collapse an \linkS4class{MTuples}-based object across strand}
\usage{
strandCollapse(x, ...)
}
\arguments{
\item{x}{An \linkS4class{MTuples} object or an object with a slot containing 
an \linkS4class{MTuples} object for which a \code{strandCollapse()} method 
has been defined.}

\item{...}{Further arguments to be passed to or from other methods}
}
\description{
Collapse an \linkS4class{MTuples}-based object across strand
}
\details{
Collapsing an \linkS4class{MTuples}-based object by strand means to flip 
tuples on the negative strand to the positive strand and shifting these by 
the appropriate offset. 

Only objects with \code{methtype(x) == "CG"} have easily defined behaviour 
when collapsing by strand. This is because 'CG' is a palindromic DNA 
sequence (it is its own reverse complement). Thus, if 
\code{methtype(x) == "CG"}, the tuples on the negative strand of \code{x} 
are flipped to the positive strand and shifted to the left by one position, 
e.g., 10 -> 9, 33 -> 32. Tuples on the positive strand are unchanged. 
Unstranded tuples are not supported and the method should throw an error if 
\code{any(strand(x) == "*" == TRUE)}. All tuples in the returned object will 
be on the positive strand.

\code{strandCollapse} is an \emph{endomorphism}, it returns an object of the 
same class as \code{x}. However, methods may collapse or aggregate 
additional slots in any manner they see fit; for an example, see 
\code{?\linkS4class{MethPat}}).
}

