% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cometh.R
\name{cometh}
\alias{cometh}
\title{Estimate within-sample, within-fragment co-methylation using 2-tuples.}
\usage{
cometh(methpat, pair_type = c("all", "strict_adjacent"), ref_loci,
  method = c("lor", "pearson"), min_cov = 5L, alternative = c("two.sided",
  "less", "greater"), conf.level = 0.95, feature, offset = 0.5)
}
\arguments{
\item{methpat}{A \code{\link{MethPat}} object containing 2-tuples.}

\item{pair_type}{A character string giving the type of pairs to be
constructed when computing correlations. One of "\code{neighbours}" or
"\code{all}", can be abbreviated. Please see the below section, "Choice of
'pair_type'", for details.}

\item{method}{A character string indicating which statistic to use to
estimate co-methylation. This must be (an abbreviation of) one of the
strings "\code{lor}" (\eqn{log_{2}(odds-ratio)}) or
"\code{pearson}" (Pearson product-moment correlation coefficient). Please
see the below section, "Choice of statistic", for details.}

\item{min_cov}{An \code{integer} specifying the minimum coverage required
in order to use a 2-tuple when estimating co-methylation.}

\item{alternative}{Indicates the alternative hypothesis and must be one of
"two.sided", "greater" or "less". You can specify just the initial letter.
"greater" corresponds to positive association, "less" to negative
association.}

\item{feature}{An optional \code{\link[GenomicRanges]{GRanges}} object with
the locations of a "genomic feature". This
\code{\link[GenomicRanges]{GRanges}} object must be disjoint (see
\code{\link[GenomicRanges]{isDisjoint}}). Please see the
below section, "Stratifying pairs by a genomic feature", for details.}

\item{offset}{A \code{numeric} vector with length 1 used when computing
M-values (default: 0.5).}
}
\description{
Estimate within-sample, within-fragment co-methylation using 2-tuples.
}
\section{Choice of 'pair_type'}{

}

\section{Choice of statistic}{

}

\section{Stratifying pairs by a genomic feature}{

}

