% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getCoverage,MethPat-method}
\alias{getCoverage,MethPat-method}
\title{Compute sequencing coverage of m-tuples.}
\usage{
\S4method{getCoverage}{MethPat}(object)
}
\arguments{
\item{object}{A \code{\link{MethPat}} object.}
}
\value{
A \code{\link[base]{matriobject}}, with the same dimensions and
dimension names as \code{object}, of sequencing coverage of each tuple in
each sample.
}
\description{
Compute sequencing coverage of m-tuples.
}

