% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MethInfo-class.R
\docType{class}
\name{methtype}
\alias{MethInfo}
\alias{MethInfo-class}
\alias{merge,MethInfo,MethInfo-method}
\alias{merge,MethInfo,NULL-method}
\alias{merge,MethInfo,missing-method}
\alias{merge,NULL,MethInfo-method}
\alias{merge,missing,MethInfo-method}
\alias{methtype}
\alias{methtype,MethInfo-method}
\alias{methtype<-}
\alias{methtype<-,MethInfo,character-method}
\alias{show,MethInfo-method}
\alias{summary,MethInfo-method}
\title{MethInfo objects}
\usage{
methtype(x)

methtype(x) <- value

\S4method{methtype}{MethInfo}(x)

MethInfo(methtype = NA_character_)

\S4method{methtype}{MethInfo,character}(x) <- value

\S4method{summary}{MethInfo}(object, ...)

\S4method{show}{MethInfo}(object)

\S4method{merge}{MethInfo,MethInfo}(x, y, ...)

\S4method{merge}{MethInfo,missing}(x, y, ...)

\S4method{merge}{MethInfo,`NULL`}(x, y, ...)

\S4method{merge}{`NULL`,MethInfo}(x, y, ...)

\S4method{merge}{missing,MethInfo}(x, y, ...)
}
\arguments{
\item{x}{The object from/on which to get/set the methylation information.}

\item{value}{A character vector of methylation types: \code{"CG"} 
(\emph{i.e.}, CpG), \code{"CHG"}, \code{"CHH"}, \code{"CN"} or some 
combination of these, e.g., \code{c("CG", "CHG")} (\code{NA_character_} is 
also allowed, but not recommended).}

\item{methtype}{A character vector of methylation types: \code{"CG"} 
(\emph{i.e.}, CpG), \code{"CHG"}, \code{"CHH"}, \code{"CN"} or some 
combination of these, e.g., \code{c("CG", "CHG")} (\code{NA_character_} is 
also allowed, but not recommended).}

\item{...}{Additional MethInfo objects.}

\item{y, object}{A MethInfo object.}
}
\description{
A \code{MethInfo} object is an object that contains basic 
information about a set of methylation loci. Currently, the only attributes 
are the type or context of the methylation loci, but more attributes may 
be added in the future as the need arises.
}
\details{
Typically \code{MethInfo} objects are not used directly but are part of 
higher level objects. Those higher level objects will generally provide a 
\code{methinfo} accessor for getting/setting their \code{MethInfo} component.
}
\section{Constructor}{

\describe{
\item{}{
 \code{MethInfo(methtype)}: Creates a \code{MethInfo} object.
 }
}
}

\section{Accessor methods}{

In the code snippets below, \code{x} is a \code{MethInfo} object.
 \describe{
   \item{}{
     \code{methtype(x)}, \code{methtype(x) <- value}:
     Get/set the methylation type of \code{x}. \code{value} must be a 
     character vector: \code{"CG"} (\emph{i.e.}, CpG), \code{"CHG"}, 
     \code{"CHH"}, \code{"CN"} or some combination of these, e.g., 
     \code{c("CG", "CHG")} (\code{NA_character_} is also allowed, 
     although not recommended).
   }
 }
}

\section{Combining MethInfo objects}{

There is no \code{c} method for \code{MethInfo} objects. Rather, a 
\code{merge} method is provided. 

In the code snippet below, \code{x} and \code{y} are \code{MethInfo} objects. 
 \describe{
   \item{}{
     \code{merge(x, y)}:
     Merge \code{x} and \code{y} into a single \code{MethInfo} object where 
     the methylation type is the union of \code{methtype(x)} and 
     \code{methtype(y)}. If the \code{methtype} of any object is missing 
     (\code{NA_character_}) then the merged \code{methtype} is also missing 
     (\code{NA_character}).
   }
}
}
\examples{
x <- MethInfo("CG")
y <- MethInfo(c("CHG", "CG"))

# Pretty-prints methylation type as "CG/CHG"
y 

# Returns the methylation type as a character vector
methtype(y)

# Merge two MethInfo objects
merge(x, y)
 
}
\author{
Peter Hickey
}

