% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{methinfo}
\alias{methinfo}
\alias{methinfo<-}
\alias{methtype}
\alias{methtype<-}
\title{Accessing/modifying methylation information}
\usage{
methinfo(x)

methinfo(x) <- value

methtype(x)

methtype(x) <- value
}
\arguments{
\item{x}{The object from/on which to get/set the methylation information.}

\item{value}{Typically a \linkS4class{MethInfo} object for the 
\code{methinfo} setter.}
}
\description{
A set of generic functions for getting/setting/modifying the 
methylation information stored in an object.
}
\details{
The \linkS4class{MethInfo} class plays a central role for the 
functions described in this man page because:
\itemize{
 \item All these functions work on a \linkS4class{MethInfo} object.
 \item Default \code{methtype} getters and setters are provided. By default, 
 \code{methtype(x)} does \code{methtype{methinfo(x)}}. So any class with a 
 \code{methinfo} getter will have all the above getters work out-of-the-box. 
}

If, in addition, the class defines a \code{methinfo} setter, then all the 
corresponding setters also work out-of-the-box.

Examples of containers that have a \code{methinfo} getter and setter: 
\linkS4class{MTuples}, \linkS4class{MTuplesList}, and \linkS4class{MethPat} 
classes.
}
\examples{
## TODO
}
\author{
Peter Hickey
}

