% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MethylationTuples.R
\docType{package}
\name{MethylationTuples-package}
\alias{MethylationTuples-package}
\title{Tools for analysing methylation patterns at genomic tuples.}
\description{
\pkg{MethylationTuples} provides tools for analysing, managing and
visualising methylation patterns at genomic tuples. These include analyses
of co-methylation and epipolymorphism.
}
\details{
Please refer to the vignettes to see how to use the \pkg{MethylationTuples}
package.
}

