% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTuples-class.R
\docType{class}
\name{MTuples-class}
\alias{MTuples}
\alias{MTuples-class}
\alias{c,MTuples-method}
\alias{methinfo,MTuples-method}
\alias{methinfo<-,MTuples,MethInfo-method}
\alias{methtype,MTuples-method}
\alias{methtype<-,MTuples,character-method}
\alias{show,MTuples-method}
\title{MTuples objects}
\usage{
MTuples(gtuples = GTuples(), methinfo = MethInfo())

\S4method{show}{MTuples}(object)

\S4method{c}{MTuples}(x, ..., ignore.mcols = FALSE, recursive = FALSE)

\S4method{methinfo}{MTuples}(x)

\S4method{methtype}{MTuples}(x)

\S4method{methinfo}{MTuples,MethInfo}(x) <- value

\S4method{methtype}{MTuples,character}(x) <- value
}
\arguments{
\item{gtuples}{A \linkS4class{GTuples} object containing the positions of 
the methylation loci as genomic tuples.}

\item{methinfo}{A \linkS4class{MethInfo} object containing information 
about the methylation loci present in the \code{MTuples} object.}

\item{x, object}{An MTuples object.}

\item{...}{Additional MTuples objects.}

\item{ignore.mcols, recursive}{See 'Splitting and Combining' section of 
\code{?}\linkS4class{GTuples}.}

\item{value}{A replacement object of the appropriate class.}
}
\description{
The \code{MTuples} class is a container for the genomic locations 
of methylation loci in a genome.
}
\details{
The \code{MTuples} class extends the \linkS4class{GTuples} class by adding 
the \code{methinfo} slot, which records the information about the type of 
methylation loci that are stored in the MTuplesList object.
}
\examples{
## TODO

}
\author{
Peter Hickey
}
\seealso{
\code{\link{findMTuples}} to find genomic tuples of methylation 
loci in a \linkS4class{BSgenome} object, and \linkS4class{GTuples} for the 
class from which MTuples inherits.
}

