% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTuples-class.R
\docType{class}
\name{MTuples-class}
\alias{MTuples-class}
\alias{c,MTuples-method}
\alias{methinfo,MTuples-method}
\alias{methinfo<-,MTuples,MethInfo-method}
\alias{methtype,MTuples-method}
\alias{methtype<-,MTuples,character-method}
\alias{show,MTuples-method}
\title{MTuples objects}
\usage{
\S4method{show}{MTuples}(object)

\S4method{c}{MTuples}(x, ..., ignore.mcols = FALSE, recursive = FALSE)

\S4method{methinfo}{MTuples}(x)

\S4method{methtype}{MTuples}(x)

\S4method{methinfo}{MTuples,MethInfo}(x) <- value

\S4method{methtype}{MTuples,character}(x) <- value
}
\arguments{
\item{object}{An MTuples object.}

\item{x}{An MTuples object.}

\item{...}{Additional MTuples objects.}

\item{ignore.mcols, recursive}{See 'Splitting and Combining' section of
\code{?}\linkS4class{GTuples}.}

\item{value}{A character vector of methylation types: \code{"CG"} 
(\emph{i.e.}, CpG), \code{"CHG"}, \code{"CHH"}, \code{"CN"} or some 
combination of these, e.g., \code{c("CG", "CHG")} (\code{NA_character_} is 
also allowed, but not recommended).}
}
\description{
The \code{MTuples} class is a container for the genomic locations 
of methylation loci in a genome.
}
\details{
The \code{MTuples} class extends the \linkS4class{GTuples} class by adding 
the \code{methinfo} slot, which records the information about the type of 
methylation loci that are stored in the MTuplesList object.
}

