% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MTuples-class}
\alias{MTuples-class}
\title{MTuples objects}
\usage{
MTuples(gtuples = GTuples(), methinfo = MethInfo())
}
\arguments{
\item{gtuples}{A \code{\link{GTuples}} object containing the positions of
the methylation loci as genomic tuples.}

\item{methinfo}{A \code{\link{MethInfo}} object containing information about
the methylation loci present in the \code{MTuples} object.}
}
\description{
The \code{MTuples} class is a container for the genomic locations
of methylation loci in a genome.
}
\details{
The \code{MTuples} class extends the
\code{\link[GenomicTuples]{GTuples}} class by adding the
\code{methinfo} slot, which records the information about the type of
methylation loci that are stored in the object.
}
\examples{
## TODO
}
\author{
Peter Hickey
}
\seealso{
\code{\link{findMTuples}} to find genomic tuples of methylation
loci in a \code{\link[BSgenome]{BSgenome}} object, and
\code{\link[GenomicTuples]{GTuples}} for the class from which \code{MTuples}
inherits.
}

