% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTuplesList-class.R
\docType{class}
\name{MTuplesList-class}
\alias{MTuplesList}
\alias{MTuplesList-class}
\alias{methinfo,MTuplesList-method}
\alias{methinfo<-,MTuplesList,MethInfo-method}
\alias{methtype,MTuplesList-method}
\alias{methtype<-,MTuplesList,character-method}
\alias{relistToClass,MTuples-method}
\alias{show,MTuplesList-method}
\title{MTuplesList objects}
\usage{
MTuplesList(...)

\S4method{methinfo}{MTuplesList}(x)

\S4method{methtype}{MTuplesList}(x)

\S4method{methinfo}{MTuplesList,MethInfo}(x) <- value

\S4method{methtype}{MTuplesList,character}(x) <- value

\S4method{relistToClass}{MTuples}(x)

\S4method{show}{MTuplesList}(object)
}
\arguments{
\item{...}{\linkS4class{MTuples} objects}

\item{x, object}{An MTuplesList object.}

\item{value}{A replacement object of the appropriate class.}
}
\description{
The \code{MTuplesList} class is a container for the genomic locations 
of methylation loci in a genome.
}
\details{
The \code{MTuplesList} class extends the \linkS4class{GTuplesList} class by 
adding the \code{methinfo} slot, which records the information about the 
type of methylation loci that are stored in the MTuplesList object.
}
\examples{
## TODO

}
\author{
Peter Hickey
}
\seealso{
\itemize{
 \item \linkS4class{GTuplesList} for the class from which MTuplesList 
 inherits
 \item \linkS4class{MTuples}
}
}

