% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTuplesList-class.R
\docType{class}
\name{MTuplesList-class}
\alias{MTuplesList}
\alias{MTuplesList-class}
\title{MTuplesList objects}
\usage{
MTuplesList(...)
}
\arguments{
\item{...}{\code{\link{MTuples}} objects. All must contain the same
\code{size} tuples.}
}
\description{
The \code{MTuplesList} class is a container for storing a collection of
\code{\link{MTuples}} objects. The \code{MTuplesList} class is almost
identical to the \code{\link[GenomicTuples]{GTuplesList}} on which it is
based.
}
\details{
Please see
\code{\link[GenomicTuples]{GTuplesList}} for a description of available
methods. The only additional methods are \code{methinfo} and
\code{\link{methtype}}, which are identical to their \code{\link{MTuples}}
counterparts.
}
\examples{
## TODO
}
\author{
Peter Hickey
}
\seealso{
\code{\link{MTuples}}, \code{\link[GenomicTuples]{GTuplesList}}.
}

