% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.methtuple}
\alias{read.methtuple}
\title{Read \code{.tsv} output files from \code{methtuple} software.}
\usage{
read.methtuple(files, sample_names = paste0("sample_", seq_along(files)),
  methinfo = MethInfo(), seqinfo = Seqinfo(), verbose = FALSE,
  bpparam = bpparam())
}
\arguments{
\item{files}{The \code{.tsv} files created by \code{comethylation}. These
files may be compressed with gzip or bzip2 (not yet implemented). All files
must contain the same sized m-tuples.}

\item{sample_names}{The sample names of each file. Must be unique.}

\item{methinfo}{A \code{\link{MethInfo}} object containing information about
the the methylation loci in the \code{files}. This should be the minimal
\code{MethInfo} object necessary to describe all methylation loci in all
\code{files}. For example, if reading in two files, one with \code{CG} and
one with \code{CHG} methylation, then the \code{methinfo} should be
\code{MethInfo(c('CG', 'CHG'))}. \strong{NB:} It is generally recommended
that you only store samples with the same \code{methinfo} in each
\code{MethPat} object.}

\item{seqinfo}{A \code{\link[GenomeInfoDb]{Seqinfo}} object containing
information about the reference genome of the samples.}

\item{verbose}{A \code{\link{logical(1)}} indicating whether messages about
the reading of the data (with \code{data.table::\link[data.table]{fread}})
and about data coercion during construction should be printed.}

\item{bpparam}{A \code{\link[BiocParallel]{bpparam}} object specifying the
parallelisation strategy, if any. See below for a discussion of
parallelisation options available with \code{read.methtuple}.}
}
\value{
A \code{\link{MethPat}} object
}
\description{
Read the \code{.tsv} output files from \code{methtuple} and construct a single
\code{\link{MethPat}} object.  \code{methtuple}
(\url{www.github.com/PeteHaitch/methtuple}) is Python software to extract
methylation patterns at methylation loci from \code{BAM} files.
All files should contain the same size m-tuples, e.g., all 2-tuples. All
files should be mapped against the same reference genome, which is supplied
as the \code{seqinfo} argument.
}
\note{
The compression of \code{files} is determined by the file extensions:
\code{.gz} for compressed with gzip and \code{.bz2} for compressed with
bzip2. In all other cases the file is assumed to be uncompressed.

Files will be decompressed to a temporary directory.
}
\section{Parallelisation}{

Parallelisation of \code{read.methtuple} is partially supported. Files
may be decompressed in parallel but not read-in in parallel.
Parallelisation uses the \pkg{BiocParallel} package. By default this uses a
\code{\link[BiocParallel]{MulticoreParam()}} instance or the user’s
preferred back-end if they have used \code{\link[BiocParallel]{register}}.
Please consult the \pkg{BiocParallel} documentation for details on
registering a parallel backend and parallelisation support available on
different operating systems.
}
\examples{
## TODO
}
\seealso{
\code{\link{MethPat}}
}

